requirements.txt
# Core dependencies
anthropic>=0.18.0
openai>=1.12.0
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0
convokit
